refactor(tuirealm): move "Props" to its own file
